+++ tests/test_ddp.py
@@ -1,7 +1,7 @@
 import time
 from pathlib import Path
 
-from waffle_hub import TaskType
+from waffle_dough.type.task_type import TaskType
 from waffle_hub.dataset import Dataset
 from waffle_hub.hub import Hub
 from waffle_hub.schema.result import TrainResult
