+++ waffle_hub/__init__.py
@@ -73,20 +73,14 @@ EXPORT_MAP = CaseInsensitiveDict(
 BACKEND_MAP = CaseInsensitiveDict(
     {
         BackendType.ULTRALYTICS: {
-            "import_path": "waffle_hub.hub.adapter.ultralytics",
-            "class_name": "UltralyticsHub",
             "adapter_import_path": "waffle_hub.hub.train.adapter.ultralytics.ultralytics",
             "adapter_class_name": "UltralyticsManager",
         },
         BackendType.AUTOCARE_DLT: {
-            "import_path": "waffle_hub.hub.adapter.autocare_dlt",
-            "class_name": "AutocareDLTHub",
             "adapter_import_path": "waffle_hub.hub.train.adapter.autocare_dlt.autocare_dlt",
             "adapter_class_name": "AutocareDltManager",
         },
         BackendType.TRANSFORMERS: {
-            "import_path": "waffle_hub.hub.adapter.transformers",
-            "class_name": "TransformersHub",
             "adapter_import_path": "waffle_hub.hub.train.adapter.transformers.transformers",
             "adapter_class_name": "TransformersManager",
         },
